+++ Engine/Controls/OptionsControl.xaml.cs
@@ -1,4 +1,5 @@
-﻿using JocysCom.ClassLibrary.Controls;
+﻿using JocysCom.ClassLibrary;
+using JocysCom.ClassLibrary.Controls;
 using System.Windows;
 using System.Windows.Controls;
 
@@ -13,6 +14,21 @@ namespace JocysCom.VS.AiCompanion.Engine.Controls
 			if (ControlsHelper.IsDesignMode(this))
 				return;
 			SettingsFolderTextBox.Text = Global.AppData.XmlFile.Directory.FullName;
+			UpdateMicrosoftControls();
+			Global.AppSettings.PropertyChanged += AppSettings_PropertyChanged;
+		}
+
+		private async void AppSettings_PropertyChanged(object sender, System.ComponentModel.PropertyChangedEventArgs e)
+		{
+			if (e.PropertyName == nameof(AppData.EnableMicrosoftAccount))
+				await Helper.Delay(UpdateMicrosoftControls);
+		}
+
+		void UpdateMicrosoftControls()
+		{
+			MicrosoftAccountsTabItem.Visibility = Global.AppSettings.EnableMicrosoftAccount
+				? Visibility.Visible
+				: Visibility.Collapsed;
 		}
 
 		private void OpenButton_Click(object sender, RoutedEventArgs e)
@@ -28,7 +44,7 @@ namespace JocysCom.VS.AiCompanion.Engine.Controls
 			if (ControlsHelper.AllowLoad(this))
 			{
 				AppHelper.InitHelp(this);
-				UiPresetsManager.InitControl(this, true);
+				UiPresetsManager.InitControl(this, true, new FrameworkElement[] { MicrosoftAccountsTabItem });
 			}
 		}
 	}